backend/requirements.txt
Flask==2.2.2
python-dotenv==1.0.0
requests==2.28.1
Pillow>=10.0.0
spotipy==2.19.0
Flask-Cors==3.0.10
